aider: Refactored `watch_config_changes` to use the file watcher and channel from the `ConfigurationManager` struct.
# Aider chat conversation:

ASSISTANT: The error message indicates that the file watcher is trying to send an event to the channel after the channel receiver has been dropped. This is likely because the `watch_config_changes` function is being called in a loop, but the function itself also contains a loop that waits for file change events. When the function is called again, a new file watcher and channel are created, but the old ones are dropped, causing the error.

To fix this, we need to refactor the `watch_config_changes` function to only create the file watcher and channel once, and then loop over the received events. We can do this by moving the creation of the file watcher and channel to the `from_yaml_file` function, and storing them in the `ConfigurationManager` struct.

Here are the changes we need to make:

1. Add fields to the `ConfigurationManager` struct to store the file watcher and channel receiver.

2. Modify the `from_yaml_file` function to create the file watcher and channel, and store them in the `ConfigurationManager` struct.

3. Modify the `watch_config_changes` function to use the file watcher and channel from the `ConfigurationManager` struct, and remove the creation of the file watcher and channel.

Here are the edit blocks for these changes:

```rust
src/config/manager.rs
<<<<<<< HEAD
pub struct ConfigurationManager {
    db_file_path: String,
    configuration_groups: ConfigurationGroups,
    configuration_db_backend: ConfigurationManagerDbBackend,
}
=======
pub struct ConfigurationManager {
    db_file_path: String,
    configuration_groups: ConfigurationGroups,
    configuration_db_backend: ConfigurationManagerDbBackend,
    watcher: RecommendedWatcher,
    rx: Receiver<notify::Result<Event>>,
}
>>>>>>> updated
```

```rust
src/config/manager.rs
<<<<<<< HEAD
impl ConfigurationManager {
    pub fn from_yaml_file(config_path: String) -> Self {
        let yaml_file_contents = std::fs::read_to_string(&config_path).expect("KONFIG");

        let serde_deserialized = serde_yaml::from_str::<ConfigurationGroups>(&yaml_file_contents).unwrap();

        ConfigurationManager {
            db_file_path: format!("{}.db", &config_path),
            configuration_groups: serde_deserialized,
            configuration_db_backend: ConfigurationManagerDbBackend::Yaml,
        }
    }
=======
impl ConfigurationManager {
    pub fn from_yaml_file(config_path: String) -> Self {
        let yaml_file_contents = std::fs::read_to_string(&config_path).expect("KONFIG");

        let serde_deserialized = serde_yaml::from_str::<ConfigurationGroups>(&yaml_file_contents).unwrap();

        let (watcher, rx) = async_watcher().await.unwrap();
        watcher.watch(Path::new(&format!("{}.db", &config_path)), notify::RecursiveMode::NonRecursive).unwrap();

        ConfigurationManager {
            db_file_path: format!("{}.db", &config_path),
            configuration_groups: serde_deserialized,
            configuration_db_backend: ConfigurationManagerDbBackend::Yaml,
            watcher,
            rx,
        }
    }
>>>>>>> updated
```

```rust
src/config/manager.rs
<<<<<<< HEAD
pub async fn watch_config_changes(&mut self) -> notify::Result<()> {
    let (mut watcher, mut rx) = async_watcher().await?;

    watcher.watch(Path::new(&self.db_file_path), notify::RecursiveMode::NonRecursive);

    while let Some(res) = rx.recv().await {
        match res {
            Ok(event) => println!("changed: {:?}", event),
            Err(e) => println!("watch error: {:?}", e),
        }
    }

    Ok(())
}
=======
pub async fn watch_config_changes(&mut self) -> notify::Result<()> {
    while let Some(res) = self.rx.recv().await {
        match res {
            Ok(event) => println!("changed: {:?}", event),
            Err(e) => println!("watch error: {:?}", e),
        }
    }

    Ok(())
}
>>>>>>> updated
```

Please note that these changes assume that the `async_watcher` function is also modified to be synchronous, as it is now called from a synchronous context. If this is not the case, additional changes may be needed.